Fixed actions not working as intended after shuffle
Fixed reference to outer this in actions command
relocatted sendCommand to actions.js
renamed Helo sendCommand to sendRequest to avoid confusion
